backgrounds was rewritten and now works, frame cycle needs to be fixed but everything else works according to plan

+++ src/backgrounds.h
@@ -6,7 +6,7 @@
 
 typedef struct Background_s
 {
-	Uint8		alive;
+	int			alive;
 	Sprite		*sprite;
 
 	Vect2d		mainPosition, chaserPosition;
